Added the latest version of CxImage (6.0.0) to CamStudio. Successful build on all projects.
From the original CxImage, I converted all the projects formats from VC++6 to VS 2008, and removed the demo, demoDLL, CxImageMfcDll and CxImageCrtDll projects from the original source since we don't need them here - the first two are for a program demonstrating the library, and the other two are for dlls (we're using CxImage as a static library).

I also added the mng, jbig and libdcr projects from CxImage to the CamStudio solution (they're required).