Modified to show the approximate running filesize in the GUI.
Based on about 15 test of various file sizes, the final saved .avi is usually a few tenths or hundredths of a percent larger than the running filesize.  This may be due to header information or something else I'm not accounting for, but this approximation is good enough for the casual user.

Next step will be to add a GUI to automatically stop recording when it reaches a certain file size.