============================== README-libconfig4Camstudio.txt ==============================
The library libconfig is not included as source code in the Camstudio source code tree because it is an third party project that is ongoing in development.
To be able to compile Camstudio you are adbvised to download and to use libconfig from:

   To be defined ....

As we can not test Camstudio with all sequential releases of libconfig you are advised to download the releases we tested and which proofed to work.
In Oct 20101 libconfig release 1.4.6 proofed to compile and to work smoothly in cooperation with Camstudio.

File CamStudio.sln is prepared to work together with libconfig.
Once you have download libconfig and put all resources is libconfig folder the main Camstudio.sln will be able to compile all and to build all the binaries.

If you have build Camstudio succesfully with other libconfig releases just let us know.

Best regards,
Camstudio development team.