Minor cleanup of the cursor annotation ring garbage collection code.
Codacy has a hard time understanding the code as it seems. The fact that
I have an assignment in the if before the condition throws it off track.
There is nothing I can do or will do about it. You should not change your
code to make the analytical tools happy. You should understand the code
that is written and recognize the false positive.